LayoutTests/fast/table/resize-table-binding-cell.html

<!DOCTYPE html>
<html>
<head>
    <style>
    td
    {
        width: 50px;
        height: 40px;
    }
    table.big td.binding
    {
        height: 80px;
    }

    table td.filler
    {
        height: auto;
    }
    </style>
    <script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
    <div style="height: 300px">
        <table border="1" style="height: 100%">
            <tr id="firstRow">
            <td class="binding"></td>
            <td></td>
            </tr>
            <tr>
            <td class="filler" ></td>
            <td class="filler" ></td>
            </tr>
        </table>
    </div>
    <script>
    var bigRows = false;

    function toggleBigRows()
    {
        var table = document.querySelector("table")
        bigRows = !bigRows;
        if (bigRows)
            table.classList.add("big");
        else
            table.classList.remove("big");
    }

    description("Regression(99212): table rows get incorrect height after changing some cells' height<br>https://bugs.webkit.org/show_bug.cgi?id=74303");
    firstRow = document.getElementById("firstRow");
    // Original value.
    shouldBe("getComputedStyle(firstRow, '').getPropertyValue('height')", "'44px'");

    toggleBigRows();
    shouldBe("getComputedStyle(firstRow, '').getPropertyValue('height')", "'84px'");
    toggleBigRows();
    shouldBe("getComputedStyle(firstRow, '').getPropertyValue('height')", "'44px'");
    </script>
    <script src="../js/resources/js-test-post.js"></script>
</body>
</html>
